requirements.txt
customtkinter>=5.2.0
psutil>=5.9.0
pystray>=0.19.0
Pillow>=10.0.0
pywin32>=306
pyyaml>=6.0
requests>=2.31.0
keyboard>=0.13.5
pygments>=2.17.0
pyinstaller>=6.0
